+++ OggConverter/Program.cs
@@ -15,7 +15,17 @@ namespace OggConverter
             Application.SetCompatibleTextRenderingDefault(false);
             if (args.Length > 0)
             {
-                Application.Run(new QuickConvert(args));
+                switch (args[0])
+                {
+                    case "wipe":
+                        Settings.WipeAll();
+                        MessageBox.Show("All your settings have been wiped. Restart the MSCMM without -wipe argument.", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        Application.Exit();
+                        break;
+                    default:
+                        Application.Run(new QuickConvert(args));
+                        break;
+                }
                 return;
             }
             Application.Run(new Form1());